+++ code/cell.module.view/src/Module.view/Module.events.ts
