+++ code/cell.module.view/src/components/ModuleView.Frame/Frame.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Subject } from 'rxjs';
 import { takeUntil, filter } from 'rxjs/operators';
 import { rx, css, CssValue, t } from '../../common';
-import { Module } from '../../Module';
+import { Module } from '../../Module.view';
 import { DebugHeader } from './Frame.DebugHeader';
 
 export type IModuleViewFrameProps = {
