+++ code/cell.module.view/src/index.ts
@@ -1,7 +1,7 @@
 import { Builder } from '@platform/cell.module';
 export { Builder };
 
-import { Module } from './Module';
+import { Module } from './Module.view';
 export { Module };
 
 export * from './components/ModuleView';
