+++ code/cell.types/src/types.Module/index.ts
@@ -1,3 +1,5 @@
 export * from './types.module';
 export * from './types.module.view';
+
 export * from './types.builder';
+export * from './types.builder.treeview';
