+++ code/cell.types/src/types.Module/types.builder.treeview.ts
@@ -1,0 +1,11 @@
+
+
+/**
+ * Builder API for a treeview node.
+ * Generics:
+ *    <P> Parent builder type
+ */
+export type ITreeviewNodeBuilder<P> = {
+  parent(): P;
+  label(value: string): ITreeviewNodeBuilder<P>;
+};
